requirements.txt
flask
opencv-python
numpy
pycuda
tensorrt
python-dotenv
